Generate the post model we start by generating the `Post` model by running ```bin/rails g model post title:string body:text``` in the terminal. This generates for us the model file and the migration file as well.
We chose `string` for title because we expect it to have fewer characters than 255 characters while we expect the body of the post to have much more charcters.

After generating the model we will need to migrate the database which we can do by running ```bin/rake db:migrate``` in the terminal while we're still in the root folder of our Rails project. This will create the table in the database.

You will notice a print out in the termial like this:
```
== 20150522143456 CreatePosts: migrating ======================================
-- create_table(:posts)
   -> 0.0017s
== 20150522143456 CreatePosts: migrated (0.0019s) =============================
```

You will also notice that the file `db/schema.rb` is modified to reflect the current state of the database. You should not modify this file manually, the migrations will take care of modifying this file.

Note: if you haven't initialized your Rails project with a `-T` option you should also see a test file getting generated as well. This is a topic for another cast.