Generate the Posts Controller
In order to make use of the model from a web page we will need to generate a controller. The convention in Rails is to make the controller the English-pluralized version of the model. So in our case our controller class should be called `PostsController`. We generate the controller by running `bin/rails g controller posts` in the terminal.

This will generate few things:
1- the `app/controllers/posts_controller.rb` file which is our controller class
2- `app/views/posts` folder where we will place our view files
3- `app/helpers/posts_helper.rb` file where we place our custom helper methods
4- `app/assets/javascripts/posts.coffee` file where we place our JavaScript code written in CoffeeScript syntax
5- `app/assets/stylesheets/posts.scss` file where we place our CSS code written in SCSS syntax

Note: if you haven't initialized your Rails project with a `-T` option you should also see test files getting generated. This is a topic for another cast.