Create the New action and view file
We start by creating the `new` action in order to have a page that shows the form to be used for creating the post. An action is really just a method in the controller so we go to our controller and we create a `new` method in there.

We then instantiate a model object in there by adding the line:
```ruby
@post = Post.new
```

We need this in order to simplify building the form by using the `form_for` helper method.

We then create the `new.html.erb` file which should be placed in the `app/views/posts` folder. Note that the name follows a Rails convention. the `new` part matches the controller action, the `.html` is the format we're using and the `.erb` part is the templating language used. `ERB` is the default Rails templating language.

After that we build the form using the `form_for`helper method and the methods that are generated with the form object `f` such as `.label`, `.text_input` and `text_area`.

It's highly recommended that you use the form helpers whenever possible. They don't just make life easier. They actually have other useful effects. For instance the `form_for` method generates an automatic authenticity token that protects against CSRF attacks.