Implement the Post Create action
After the new post form is submitted we will need to capture the parameters that are provided by the user to create a `Post` record in the database.

We start by adding the `create` method in the `PostsController`.

Then we will need to capture the parameters from user, we do so by defining `post_params`:
```ruby
post_params = params.require(:post).permit(:title, :body)
```
This is a feature introduced in Rails 4 called Strong parameters which requires the developer by default to be explicit about the allowed parameters for security reasons.

We then instantiate a new `Post` object:
```ruby
@post = Post.new(post_params)
```
We then save it:
```ruby
@post.save
```
And redirect to a new post page:
```ruby
redirect_to posts_path, notice: "Post Created"
```
Note that the posts_path takes us to `PostsController` with `index` action which doesn't exist at the moment.