Handle unsuccessful saving of a post
We start by adding validations to the `Post` model:
```ruby
validates :title, presence: true
```
This will prevent the model from persisting the database if the `title` attribute is not provided.

We then alter our code in the `create` action in the `PostsController` so that if the record is saved correctly, which we know from the `save` method returning true, we redirect to the `posts_path` as we did before. If the record doesn't save correctly, which we know from the `save`  method returning false, we render the new template which is the same template we used in for the `new` action. This way the form will display again for the user to correct the error.

Finally, we add error display code in the `new.html.erb` template so the user knows what went wrong. We use the `.errors` method which gives us an object that has the errors resulted from saving the record.