create the index action in the posts controller
Start by creating the index action in the `PostsController`. It's a convention in Rails to use this method for listing a group of records.

All we need to do here is to define an instance variable called `@posts` which contains all the `Post` records in the database. We will need this for our view files.