create the index.html.erb file
In order to display all the posts on the web page, we will need to have a view file. As per Rails' conventions, we call our file `index.html.erb` and we place it in the `views/posts` folder.

The file loops through the `@posts` variable which we defined in the `PostsController` inside the `index` action (method).

Inside the loop we display HTML code for each one of those posts.