create the show.html.erb template
Now we need to create the `show.html.erb` file which we put in the `app/views/posts/` folder as per Rails conventions.

Inside the template we can make use of the `@post` which is accessible in this template file as we're rendering the `show.html.erb` within the `show` action in the `PostsController`.