make the title in the index page link to the show page
We now convert the title listed in the `index` page which lists all the posts to be a link to the post show page. We do so by using the `link_to` Rails helper method which generates an `<a>` tag. We pass it the first argument which is the anchor text. In our case the anchor text is just the post title. The second argument of the `link_to` method is the url path. For that we use `post_path(post)` which is genertaed from the `resources :posts` line in `routes.rb` file.