Add the edit ERB template
We then add the `edit.html.erb` file in `/app/views/posts/` folder. This file looks very similar to the `new.html.erb` where it contains a form that is used to update the record.

Please note that the Rails `form_for` method looks at the first passed argument which is `@post` in this case. If this argument is persisted to the database then the form will submit a `patch` request and it will submit to `/posts/:id` url where id is the url of the `@post` object.

Also, Rails will auto-populate the fields with the values assigned in the `@post` object. So the text box generated in the will be pre-populated with what's currently in the database.