Refactor - Part 1
Let's improve our code by refactoring. Refactoring is changing the organizing of the code to make more readable, more extensible and with less duplication without actually changing the functionality of the code itself.

We first notice that the the `new.html.erb` and the `edit.html.erb` are very similar. We can refactor this by moving the duplicate code to another file. We call this file `_form.html.erb`. This file is called a partial in Rails.

Then we can render this from from the `new.html.erb` and `edit.html.erb` by using the `render` Rails helper method. We can simply do `<%= render "form" %>` because we named the file `_form.html.erb` and we've put it in the same directory.