Refactor - Part 2
We notice in the `PostsController` that there is duplication in code. Multiple methods, namely the `edit`, `update` and `show` have the piece of code `@post = Post.find params[:id]`.

We can make use of the `before_action` method in Rails controllers to overcome that issue. So we define a private method called `find_post` that contains that line of code.

We then call that method  before each one of those actions by passing the `only` attribute to the `before_action` method. This way the `@post` method will be defined before each one of those method calls and it will be accessible within these methods because it's defined as an instance variable.