refactor - Part 3
We also notice that in the `PostsController` the way we fetch and permit the params for the `create` and `update` actions is the same.

In Ruby you access a local method or a variable the same way. So we can refactor this piece of code by having a method called `post_params` that contains that code that fetches and permits the needed parameters.

We can make that method private because we don't need to access it as an action.