refactor - Part 4
We can simplify redirecting to the show page by using `redirect_to @post` instead of doing `redirect_to post_path(@post)`. redirect_to will automatically redirect to the show page with url `/posts/:id` if you pass it an `ActiveRecord` object. It will use the is the `id` of that object in the url.