implement the destroy action
We start implement the destroy action by defining the method `destroy` in the `PostsController`.

We will need to have access to the `@post` method so we can add the `:destroy` symbol to the array passed for the `:only` option in the `before_filter` method at the top of our controller.

We can simply destroy a record by calling the `destry` method on the `@post` object.

We then redirect the user to the post listing page with a notice the says `Post Deleted`.