add the delete link
We add the deleted link on the show page by using the `link_to` method in the `show.html.erb` template.

We must provide the link with the `method: :delete` option because to access the `destroy` method in our controller we will need to make a `delete` request to `/posts/:id`.

We can optionally pass in the `data: {confirm: "Are you sure?"}` option in order to prompt the user with a confirm dialog before acually making the `delete` request.