Add a user model
To build user authentication system we start by creating a `user` model. We do so by running the Rails generator: `bin/rails g model user first_name last_name email password_digest`. They are all strings so we didn't have to specify type. Note that we must call the field password_digest to make use of the Rails build-in user authentication feature.

Before running `bin/rake db:migrate` we can optionally add an index on the email field to the migration. This because we will be doing searches on the users table by the email of the user. Having an index on the email field speeds up such queries. We can optionally add `unique: true` option to the index which enforces uniqueness at the database level.

We then add the email validation to the user model. We want to make sure that it's present, unique and has a valid email format.