Add has_secure_password method
We add the `has_secure_password` method to the `user.rb` file. This basically gives us few things:
* Adds attribute accessors to the user model for password and password_confirmtion
* Validates presence of password and password matching with password_confirmtion if present
* Encrypts the password and stores in in password_digest using the `bcrypt` gem

For more details on this method checkout this link [http://api.rubyonrails.org/classes/ActiveModel/SecurePassword/ClassMethods.html#method-i-has_secure_password]

In order for this method to work we have to have the `bcrypt` gem added to our Gemfile. The gem is commented out by default with a new Rails project so we just uncomment that line and run `bundle install` in our terminal.