Implement the users controller actions
We now implement the new and create actions for the users controller. We first make sure to have the routes set up by adding `resources :users, only: [:new, :create]`. Then we implement the new and create actions as standard ones with Rails CRUD. After that we implement the `new.html.erb`template.

It's generally a good idea to make sure that we have the `root_path` setup. In this case we make it go to the `index` action in the `posts` controller.

The only difference from a standard CRUD is that after the users successfully create their user record we add `sessions[:user_id] = @user.id`. This way we authenticate the user to our website using the session upon successful account creation.