Implement the sessions controller
We implement the sessions controller as a standard Rails controller with two actions new and create. We utilize the `@user.authenticate` method which comes with the `has_secure_password` method. This method takes it a password and hashes it the same way it was hashed the first time and compares the outcome with password_digest. If the outcome is the same it means that the user has entered the correct password so the method returns `true`. If the password is not the same then the method will return `false`. This is because we can't go from password_digest to a clear-text password. We can only go one way with Hashing algorithms.

If the user is found an authenticated properly then we authenticate the user by setting the session variable as in:
```ruby
sessions[:user_id] = @user.id
```
We user `form_tag` for the session `new` view template because we're not really creating a model and don't have Session model that stores to the database.