Implement the sessions destroy action
We start by adding the `destroy` routes in our `routes.rb` file. We can make as a standard destroy route but then we will need to have an `:id` part of the route. This is not really needed or recommended in the case of the user. We have the user id stored in the session so we don't really need to have it part of the url. Also it's generally advisable to hide the user id whenever possible for security reasons.

We can accomplish this by adding the `destroy` route with `on: :collection` option. Which makes the route without `:id` or `:user_id` in it.

In the controller we implement the destroy action which simply sets the `sessions[:user_id]` to `nil` making our user signed out.