Implement controller and helper methods and integrate them
We not implement few controller methods that will help us utilize user authentication within our application. We put those methods in the `ApplicationController` because we want them to be available in all of our controllers.

We make the `current_user` and `user_signed_in?` helper methods by adding `helper_method :current_user` and `helper_method user_signed_in?` because we need to access those in the view files as well as controller files.

To make sure that a user is authenticated before accessing a specific controller action we simply add a `before_action` as in:
```ruby
before_action :authenticate_user!, except: [:index, :show]
```
Which will redirect the user to the sign in page if they are not authenticated.