Set up the associations
In order to set up a many-to-many association in Rails, it's best to start by having a third model to join the two models together. In this case we want to implement a liking feature so let's call the model `like`.

The `like` model has to include two references: one for the `user` model and the other for `post` model.

After running the migration, we set up the associations using `has_many` but we have to set up two in every model. For instance for the `User` model we setup one `has_many` to the `likes` and the other one to the `posts` using the `through` option for `has_many`.