Set up the likes controller
In order for a user to like a post we need to create a likes controller to handle the create action. We start by creating a likes controller using the Rails generator and we make the `resources :likes` nested within the `resources :posts`. This is because we want to like a specific post so it's best to have the post id part of the URL.

In the likes controller we start by implementing the create action. The `create` action needs to create a `like` record in the database. This record has to be associated with the `post` whose id is passed in through `params[:post_id]` and the user record which we can get from `current_user` method we implemented earlier in the `ApplicationController`. We redirect back to the `post` show page whether the like saves correctly or not showing different flash messages.

For the `destroy` action simply finds the like using `current_user.likes` so that only the user who owns a like can delete it. Then we call the `destroy` method on it. We then redirect to the post show page.