Add links to the show page
We now add links to the show page. The `like` link requires a `post` request to the `post_likes_path` and this only requires the `@question` variable.

The `un-like` button requires both `@question` and `@like` so we define the `@like` variable in the `PostsController` within the `show` action. For that we implement the `like_for` method that helps us get the like object that associates the `@question` with `current_user`.