Configure mail settings
We start setting up our Email settings with Rails by creating a file called `app_keys.rb` in the `config/initializers` folder. This will contain the username and password for the mail server we're going to use to send email. Because this file contains sensitive data, it should not go in our source control so we make sure that it's in the `.gitignore` file. The content of the `app_key.rb` should be something like:
```ruby
ENV["email_user_name"] = "YOUR_EMAIL_USERNAME_HERE"
ENV["email_password"]  = "YOUR_EMAIL_PASSWORD_HERE"
```
We also create a file `setup_mail.rb` which will include the actual configuration for the mail server. In our case we're connecting to a Gmail mail server.

When you deploy your app to production make sure to have the `ENV["email_user_name"]` and `ENV["email_password"]` set properly. If you're deploying to Heroku, you can use their web interface to set those variables. If you're deploying to EC2 or to a VPS you may need to set up a way to have `app_key.rb` present there.