Implement the email method
In order to send out an email we need to define a method for every email kind we want to send. So we define a method called `notify_post_owner` in the `LikesMailer` class. We can define this method as an instance method.

For that to work we need to do two things:
- Have at least one view file in the `app/views/likes_mailer`. Either `notify_post_owner.text.erb` or `notify_post_owner.html.erb`. It's best to have both so you send an email that is supported by html and non-html email clients.
- Call the `mail` method in the `notify_post_owner` method. You must at least give the method a `to` option to send the email.

After the `like` objects saves successfully in the `LikesContrller` we send an email to the question owner by calling: `LikesMailer.notify_post_owner(like).deliver_now`. Calling `deliver_now` makes it send the email in the foreground.

Note that inside the `app/views/likes_mailer/notify_post_owner.html.erb` we used `question_url` instead of `question_path`. This is because when someone click on a link in an email they are accessing your website from another domain or application so the `path` version won't work. For the URL in the mailer to work we have to tell it the domain to use. We did so in the `config/development.rb` file with the line:
```ruby
config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
```